Add CallNotification class and related commands
